requirements.txt
Django==1.8.5
django-appconf==1.0.1
django-ckeditor==5.0.3
django-contact-form==1.1
django-favicon==0.1.2
django-imagekit==3.2.7
pilkit==1.1.13
Pillow==3.0.0
six==1.10.0
